Refactor: standardize `User` and `Translatable` model paths to lowercase for consistency ♻️
- Renamed `User.ts` to `user.ts` and updated related imports.
- Renamed `Translatable.ts` to `translatable.ts` and updated references across the codebase.
- Adjusted stubs and service provider mappings to reflect the changes.
