+++ resources/js/inertia/components/datatable/table-body.tsx
@@ -1,4 +1,4 @@
-import { translate } from "@/models/Translatable";
+import { translate } from "@/models/translatable";
 import { getNestedPropertyValue } from "@/helper";
 import { TableBodyProps } from "@/components/datatable/types";
