--- resources/js/inertia/models/User.ts
@@ -1,6 +1,0 @@
-interface User {
-  name: string;
-  email: string;
-}
-
-export default User;
