+++ resources/js/inertia/models/translatable.ts
