+++ resources/js/inertia/models/user.ts
@@ -1,5 +1,7 @@
-export interface User {
+interface User {
   first_name?: string;
   last_name?: string;
   email: string;
 }
+
+export default User;
